pyproject.toml
[project]
name = "brand-performance-report"
version = "0.1"
description = ""
authors = [
    {name = "Maiara Souza"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "duckdb (>=1.1.3,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "dbt-core (>=1.9.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dbt-duckdb (>=1.9.1,<2.0.0)",
    "pandera[pandas] (<0.24.0)",
    "dagster-dbt (>=0.26.21,<0.27.0)",
    "dagster-duckdb (>=0.26.21,<0.27.0)",
    "dagster-pandera (>=0.26.21,<0.27.0)",
    "dagster-webserver (>=1.10.21,<2.0.0)",
    "dagster (>=1.10.21,<2.0.0)",
    "etltools @ file:///D:/Users/KO/OneDrive%20-%20The%20Coca-Cola%20Company/KO%20Unified%20Data%20Infrastructure/Brand%20Performance%20Report/transform/brand_performance/ingestion",
    "dagster-pipes (>=1.10.21,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"


[tool.poetry.dependencies]
etltools = {develop = true}
